% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabilog.R
\name{AIC.mabilog}
\alias{AIC.mabilog}
\title{Compute AIC for Mabilog Model}
\usage{
\method{AIC}{mabilog}(object, ..., k = 2)
}
\arguments{
\item{object}{A 'mabilog' object}

\item{...}{Additional arguments passed to logLik}

\item{k}{Penalty parameter (default: 2)}
}
\value{
AIC value
}
\description{
Computes Akaike Information Criterion for model selection
}
