% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabilog.R
\name{BIC.mabilog}
\alias{BIC.mabilog}
\title{Compute BIC for Mabilog Model}
\usage{
\method{BIC}{mabilog}(object, ...)
}
\arguments{
\item{object}{A 'mabilog' object}

\item{...}{Additional arguments passed to logLik}
}
\value{
BIC value
}
\description{
Computes Bayesian Information Criterion for model selection
}
