% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabilo_plus.R
\name{coef.mabilo_plus}
\alias{coef.mabilo_plus}
\title{Extract Model Coefficients from Mabilo Plus Model}
\usage{
\method{coef}{mabilo_plus}(object, ...)
}
\arguments{
\item{object}{A 'mabilo_plus' object}

\item{...}{Additional arguments (currently unused)}
}
\value{
Named vector containing optimal k values
}
\description{
Extracts the optimal k values from a mabilo_plus object
}
