% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magelo.R
\name{deg0.loo.llm.1D}
\alias{deg0.loo.llm.1D}
\title{Degree 0 case of Leave-One-Out (LOO) cross-validation of 1D local linear models}
\usage{
deg0.loo.llm.1D(nx, nn.i, nn.w, nn.y)
}
\arguments{
\item{nx}{The number of elements of x.}

\item{nn.i}{A matrix of indices of x NN's of xgrid.}

\item{nn.w}{A matrix of NN weights.}

\item{nn.y}{The values of y over nn.i.}
}
\value{
predictions
}
\description{
Degree 0 case of Leave-One-Out (LOO) cross-validation of 1D local linear models
}
