% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabilog.R
\name{deviance.mabilog}
\alias{deviance.mabilog}
\title{Compute Deviance for Mabilog Model}
\usage{
\method{deviance}{mabilog}(object, ...)
}
\arguments{
\item{object}{A 'mabilog' object}

\item{...}{Additional arguments (currently unused)}
}
\value{
Deviance value
}
\description{
Computes the deviance of the fitted model
}
