% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{epanechnikov.kernel}
\alias{epanechnikov.kernel}
\title{Epanechnikov kernel function}
\usage{
epanechnikov.kernel(x, bw = 1)
}
\arguments{
\item{x}{A numeric vector.}

\item{bw}{A bandwidth numeric parameter.}
}
\description{
The function equals to 1-t^2 for t in (-1,1) and 0 otherwise; t = abs(x/bw)
}
