% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwlm.R
\name{find_optimal_breakpoints}
\alias{find_optimal_breakpoints}
\title{Find Optimal Number of Breakpoints}
\usage{
find_optimal_breakpoints(
  x,
  y,
  max_breakpoints = 5,
  method = c("aic", "bic", "davies"),
  alpha = 0.05
)
}
\arguments{
\item{x}{Numeric vector containing the predictor variable values}

\item{y}{Numeric vector containing the response variable values}

\item{max_breakpoints}{Maximum number of breakpoints to consider (default = 5)}

\item{method}{Character string specifying the selection method ("aic", "bic", or "davies")}

\item{alpha}{Significance level for Davies' test (default = 0.05)}
}
\value{
A list containing the optimal number of breakpoints and model comparison results
}
\description{
Find Optimal Number of Breakpoints
}
