% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabilog.R
\name{fitted.mabilog}
\alias{fitted.mabilog}
\title{Extract Fitted Values from Mabilog Model}
\usage{
\method{fitted}{mabilog}(object, type = c("response", "link"), ...)
}
\arguments{
\item{object}{A 'mabilog' object}

\item{type}{Character string specifying which values to return:
"response" for probabilities (default), "link" for logit scale}

\item{...}{Additional arguments (currently unused)}
}
\value{
Numeric vector of fitted values
}
\description{
Extracts fitted probability values from a mabilog object
}
