% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magelo.R
\name{get.BB.gpredictions}
\alias{get.BB.gpredictions}
\title{Generates Bayesian bootstrap estimates of gpredictions}
\usage{
get.BB.gpredictions(
  n.BB,
  nn.i,
  nn.x,
  nn.y,
  y.binary,
  nn.w,
  nx,
  max.K,
  degree,
  grid.nn.i,
  grid.nn.x,
  grid.nn.w,
  grid.max.K
)
}
\arguments{
\item{n.BB}{The number of Bayesian bootstrap iterations}

\item{nn.i}{A matrix of NN indices.}

\item{nn.x}{A matrix of x values over K nearest neighbors of each element of the grid, where K is determined in the parent routine.}

\item{nn.y}{A matrix of y values over K nearest neighbors of each element of the grid.}

\item{y.binary}{Set to TRUE if y values are within the interval \code{[0,1]}.}

\item{nn.w}{A matrix of NN weights.}

\item{nx}{The number of elements of x.}

\item{max.K}{A vector of \strong{indices} indicating the range where weights are not 0.}

\item{degree}{The degree of the local models.}

\item{grid.nn.i}{A matrix of grid associated NN indices.}

\item{grid.nn.x}{A matrix of grid associated x values over NNs.}

\item{grid.nn.w}{A matrix of grid associated NN weights.}

\item{grid.max.K}{A vector of grid associated max.K values.}
}
\description{
Generates Bayesian bootstrap estimates of gpredictions
}
