% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{get.bws}
\alias{get.bws}
\title{Gets bandwidths defined as radii of linear model's disks of support}
\usage{
get.bws(d, min.K, bw)
}
\arguments{
\item{d}{A numeric matrix nn.d of distances to NNs.}

\item{min.K}{The minimum number of elements of each row with weights > 0.}

\item{bw}{A bandwidth.}
}
\value{
A vector of bandwidths.
}
\description{
Finds a bandwidth for each column of d, such that \code{bws[i] = bw}
if \code{d[iminK + ir] < bw} and \code{bws[i] = d[iminK + 1 + ir] = d[minK + ir]}, otherwise.
}
