% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{get.bws.with.minK}
\alias{get.bws.with.minK}
\title{Gets bandwidths defined as radia of linear model's disks of support}
\usage{
get.bws.with.minK(d, minK, bw)
}
\arguments{
\item{d}{A numeric matrix.}

\item{minK}{A vector such that minK[i] is the required minimum number of
elements in the row of the i-th column of d with weights > 0.}

\item{bw}{A bandwidth.}
}
\value{
A vector of bandwidths.
}
\description{
The difference between this function and get.bws() is that in the former,
min.K is a constant, which is replaced by a vector minK in here. This
function is not meant to be accessible by user and serves as an R interface
to get_bws_with_minK_a() for unit testing this C function.
}
