% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magelo.R
\name{get.gpredictionss}
\alias{get.gpredictionss}
\title{Estimates gpredictions Means Matrix Over a Uniform Grid for Different bandwidths}
\usage{
get.gpredictionss(bws, nn.i, nn.d, nn.x, nn.y, y.binary, degree, min.K, xgrid)
}
\arguments{
\item{bws}{A vector of bandwidths.}

\item{nn.i}{A matrix of indices for K nearest neighbors.}

\item{nn.d}{A matrix of distances to the nearest neighbors.}

\item{nn.x}{A matrix of x values over K nearest neighbors of each grid point, where K is determined in the parent routine.}

\item{nn.y}{A matrix of y values over K nearest neighbors of each grid point.}

\item{y.binary}{Set to TRUE if y values are in the interval \code{[0,1]}.}

\item{degree}{The degree of the local models.}

\item{min.K}{The minimal number of elements in each set of nearest neighbors.}

\item{xgrid}{A vector of grid points.}
}
\value{
A matrix of the estimates of the means, gpredictions, of y over a uniform grid where the i-th column consists of gpredictions estimates using the i-th value of the 'bws' vector.
}
\description{
Given a vector, bws, of bandwidths this routine estimates a matrix of gpredictions's
where each column corresponds to the gpredictions estimate for a different bandwidth.
This is an R interface to a C routine, C_get_Eygs(), where the estimates of
gpredictions's are done. In practice, it is not intended to be used directly from R,
but rather to test the correctness of C_get_Eygs(). This routine is used in
different bandwidth optimization processes based on gpredictions characteristics (like
total absolute curvature or the number of inflection points).
}
\examples{
\dontrun{
res <- get.gpredictionss(bws, nn.i, nn.d, nn.x, nn.y, y.binary, degree, min.K, xgrid)
str(res)
}
}
