% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magelo.R
\name{lcor.1D}
\alias{lcor.1D}
\title{1D local weighed Pearson correlation model}
\usage{
lcor.1D(
  x,
  y1,
  y2,
  grid.size = 400,
  f = NULL,
  bw = NULL,
  smooth = TRUE,
  n.BB = 0,
  get.predictions.CrI = TRUE,
  level = 0.95,
  n.C.itr = 100,
  C = 6,
  stop.C.itr.on.min = TRUE,
  n.cv.folds = 10,
  n.cv.reps = 5,
  min.K = 5,
  nn.kernel = "epanechnikov",
  y1.binary = FALSE,
  cv.nNN = 3,
  verbose = FALSE
)
}
\arguments{
\item{x}{A numeric vector of a predictor variable.}

\item{y1}{A numeric vector of the first outcome variable.}

\item{y2}{A numeric vector of the second outcome variable.}

\item{grid.size}{A number of grid points; was grid.size = 10*length(x), but the
results don't seem to be different from 400 which is much faster.}

\item{f}{The proportion of the range of x that is used within a moving window
to train the model. If NULL, the optimal value of f will be found using
minimum median absolute error optimization algorithm.}

\item{bw}{A bandwidth parameter.}

\item{smooth}{Set to TRUE (default) to smooth the estimates of local weighted correlations.}

\item{n.BB}{The number of Bayesian bootstrap (BB) iterations for estimates of CI's of beta's.}

\item{get.predictions.CrI}{A logical parameter. If TRUE, BB with quantile determined by the value of 'level' will be used to determine the upper and lower limits of CI's.}

\item{level}{A confidence level.}

\item{n.C.itr}{The number of Cleveland's absolute residue based reweighting iterations for a robust estimates of mean y values.}

\item{C}{A scaling of |res| parameter changing |res| to |res|/C  before applying ae.kernel to |res|'s.}

\item{stop.C.itr.on.min}{A logical variable, if TRUE, the Cleveland's iterative reweighting stops when the maximum of the absolute values of
differences of the old and new predictions estimates are reach a local minimum.}

\item{n.cv.folds}{The number of cross-validation folds. Used only when f = NULL. Default value: 10.}

\item{n.cv.reps}{The number of repetitions of cross-validation. Used only when f = NULL. Default value: 5.}

\item{min.K}{The minimal number of x NN's that must be present in each window.}

\item{nn.kernel}{A kernel.}

\item{y1.binary}{Set to TRUE if y1 is a binary variable.}

\item{cv.nNN}{The number of nearest neighbors in interpolate_gpredictions() used to find predictions given gpredictions in the cv_deg0_ routines.}

\item{verbose}{Prints info about what is being done.}
}
\value{
A list of input parameters as well as coefficients and residues of all linear models
}
\description{
1D local weighed Pearson correlation model
}
