% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magelo.R
\name{llm.1D.beta.perms}
\alias{llm.1D.beta.perms}
\title{Estimates the coefficients of linear models in the vicinity of each grid point of radius nn.r on the permuted version}
\usage{
llm.1D.beta.perms(nn.x, nn.i, y, nn.w, max.K, degree, n.perms)
}
\arguments{
\item{nn.x}{A matrix of x values over K nearest neighbors of each element of the grid, where K is determined in the parent routine.}

\item{nn.i}{A matrix of indices of the K nearest neighbors of each element of the grid.}

\item{y}{A vector of response values.}

\item{nn.w}{A matrix of weights over K nearest neighbors of each element of the grid.}

\item{max.K}{A vector of indices indicating the range where weights are not 0.}

\item{degree}{A degree of the polynomial of x in the linear regression. The only allowed values are 1 and 2.}

\item{n.perms}{Number of permutations to perform. Must be a positive integer.}
}
\value{
A numeric vector of beta coefficients from all permutations.

NOTE: Weights must sum up to 1!
}
\description{
The max.K parameter is used to avoid looping over indices where weights are 0.
}
