% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magelo.R
\name{llm.1D.beta}
\alias{llm.1D.beta}
\title{Eestimates the coefficients of linear models in the vicinity of each grid point of radius nn.r}
\usage{
llm.1D.beta(nn.x, nn.y, nn.w, max.K, degree)
}
\arguments{
\item{nn.x}{A matrix of x values over K nearest neighbors of each element of the grid, where K is determined in the parent routine.}

\item{nn.y}{A matrix of y values over K nearest neighbors of each element of the grid.}

\item{nn.w}{A matrix of weights over K nearest neighbors of each element of the grid.}

\item{max.K}{A vector of indices indicating the range where weights are not 0.}

\item{degree}{A degree of the polynomial of x in the linear regression. The only allowed values are 1 and 2.}
}
\value{
list with components: beta, gpredictions.

NOTE: Weights must sum up to 1!
}
\description{
The max.K parameter is used to avoid looping over indices where weights are 0.
}
