% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magelo.R
\name{llm.1D.fit.and.predict.global.BB.qCrI}
\alias{llm.1D.fit.and.predict.global.BB.qCrI}
\title{Creates BB quantile-based estimates of predictions CI's}
\usage{
llm.1D.fit.and.predict.global.BB.qCrI(
  y.binary,
  nn.i,
  nn.w,
  nn.x,
  nn.y,
  max.K,
  degree,
  nx,
  n.BB,
  alpha = 0.05
)
}
\arguments{
\item{y.binary}{A logical variable. If TRUE, predicted predictions will be trimmed to the closed interval \code{[0, 1]}.}

\item{nn.i}{A matrix of the indices of K nearest neighbors of each element of the grid, where K is determined in the parent routine.}

\item{nn.w}{A matrix of weights.}

\item{nn.x}{A matrix of x values over K nearest neighbors of each element of the grid.}

\item{nn.y}{A matrix of y values over K nearest neighbors of each element of the grid.}

\item{max.K}{An array of indices indicating the range where weights are not 0. Indices < max.K at i have weights > 0.}

\item{degree}{A degree of the polynomial of x in the linear regression. The only allowed values are 1 and 2.}

\item{nx}{The number of elements of x.}

\item{n.BB}{The number of BB iterations.}

\item{alpha}{The confidence level.}
}
\value{
predictions.qCI matrix with two rows (upper and lower CI) and nx columns.
}
\description{
Creates BB quantile-based estimates of predictions CI's
}
