% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magelo.R
\name{llm.1D.fit.and.predict.global.BB}
\alias{llm.1D.fit.and.predict.global.BB}
\title{Creates Bayesian bootstrap (BB) estimates of predictions using global reweighting of the elements of x}
\usage{
llm.1D.fit.and.predict.global.BB(
  nn.i,
  nn.w,
  nn.x,
  nn.y,
  max.K,
  degree,
  nx,
  n.BB
)
}
\arguments{
\item{nn.i}{A matrix of the indices of K nearest neighbors of each element of the grid, where K is determined in the parent routine.}

\item{nn.w}{A matrix of weights.}

\item{nn.x}{A matrix of x values over K nearest neighbors of each element of the grid.}

\item{nn.y}{A matrix of y values over K nearest neighbors of each element of the grid.}

\item{max.K}{An array of indices indicating the range where weights are not 0. Indices < max.K at i have weights > 0.}

\item{degree}{A degree of the polynomial of x in the linear regression. The only allowed values are 1 and 2.}

\item{nx}{The number of elements of x.}

\item{n.BB}{The number of BB iterations.}
}
\value{
BB predictions matrix of dim nx-by-n.BB.
}
\description{
Creates Bayesian bootstrap (BB) estimates of predictions using global reweighting of the elements of x
}
