% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magelo.R
\name{llm.1D.fit.and.predict}
\alias{llm.1D.fit.and.predict}
\title{Fits 1D rllm model and generates predictions estimates}
\usage{
llm.1D.fit.and.predict(nn.i, nn.w, nn.x, nn.y, y.binary, max.K, degree, nx)
}
\arguments{
\item{nn.i}{A matrix of the indices of K nearest neighbors of each element of the grid, where K is determined in the parent routine.}

\item{nn.w}{A matrix of weights.}

\item{nn.x}{A matrix of x values over K nearest neighbors of each element of the grid.}

\item{nn.y}{A matrix of y values over K nearest neighbors of each element of the grid.}

\item{y.binary}{Set to TRUE if y's values are within the interval \code{[0,1]}.}

\item{max.K}{An array of indices indicating the range where weights are not 0. Indices < max.K at i have weights > 0.}

\item{degree}{A degree of the polynomial of x in the linear regression. The only allowed values are 1 and 2.}

\item{nx}{The number of elements of x.}
}
\value{
list with components: beta, predictions.

NOTE: Weights must sum up to 1!
}
\description{
Fits 1D rllm model and generates predictions estimates
}
