% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magelo.R
\name{llm.predict.1D}
\alias{llm.predict.1D}
\title{Compute predictions from local linear model coefficients}
\usage{
llm.predict.1D(beta, nn.i, nn.d, nn.x, nx, max.K, nn.kernel = "epanechnikov")
}
\arguments{
\item{beta}{Coefficients of the local linear model.}

\item{nn.i}{A matrix of indices of x NN's of xgrid.}

\item{nn.d}{A matrix of NN distances.}

\item{nn.x}{The values of predictor variable over NN's of xgrid.}

\item{nx}{The length of x.}

\item{max.K}{A vector of indices indicating the range where weights are not 0.}

\item{nn.kernel}{A kernel used for generating weights.}
}
\value{
A numeric vector of predictions
}
\description{
Compute predictions from local linear model coefficients
}
