% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabilog.R
\name{logLik.mabilog}
\alias{logLik.mabilog}
\title{Extract Log-Likelihood from Mabilog Model}
\usage{
\method{logLik}{mabilog}(object, ...)
}
\arguments{
\item{object}{A 'mabilog' object}

\item{...}{Additional arguments (currently unused)}
}
\value{
Log-likelihood value with attributes
}
\description{
Computes the log-likelihood for the fitted model
}
