% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magelo.R
\name{loo.llm.1D}
\alias{loo.llm.1D}
\title{Leave-One-Out (LOO) cross-validation of 1D local linear models}
\usage{
loo.llm.1D(
  x,
  y,
  grid.size = 400,
  degree = 2,
  f = 0.2,
  bw = NULL,
  min.K = 5,
  nn.kernel = "epanechnikov"
)
}
\arguments{
\item{x}{A numeric vector of a predictor variable.}

\item{y}{A numeric vector of an outcome variable.}

\item{grid.size}{A number of grid points; was grid.size = 10*length(x), but the
results don't seem to be different from 400 which is much faster.}

\item{degree}{A degree of the polynomial of x in the linear regression; 0
means weighted mean, 1 is regular linear model lm(y ~ x), and deg = d is
lm(y ~ poly(x, d)). The only allowed values are 1 and 2.}

\item{f}{The proportion of the range of x that is used within moving window to train the model.}

\item{bw}{A bandwidth parameter.}

\item{min.K}{The minimal number of x NN's that must be present in each window.}

\item{nn.kernel}{A kernel.}
}
\value{
list of parameters and residues of all linear models
}
\description{
Leave-One-Out (LOO) cross-validation of 1D local linear models
}
