% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malo-package.R
\docType{package}
\name{malo-package}
\alias{malo}
\alias{malo-package}
\title{malo: Legacy One-Dimensional Model-Averaged Local Regression}
\description{
Transitional package for legacy one-dimensional non-linear regression APIs
extracted from \pkg{gflow}.
}
\author{
\strong{Maintainer}: Pawel Gajer \email{pgajer@gmail.com}

}
\keyword{internal}
