% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magelo.R
\name{mllm.1D.fit.and.predict}
\alias{mllm.1D.fit.and.predict}
\title{Fits 1D rllm model and generates predictions estimates for each column of a matrix Y}
\usage{
mllm.1D.fit.and.predict(Y, y.binary, nn.i, nn.w, nn.x, max.K, degree)
}
\arguments{
\item{Y}{A matrix with the number of rows that is the same as the length of x that was used to construct nn.* matrices. The main application of this routine is for the case when the columns of Y are permutations some y.}

\item{y.binary}{Set to TRUE, if the values of all columns of Y are within the interval \code{[0,1]}.}

\item{nn.i}{A matrix of the indices of K nearest neighbors of each element of the grid, where K is determined in the parent routine.}

\item{nn.w}{A matrix of weights.}

\item{nn.x}{A matrix of x values over K nearest neighbors of each element of the grid.}

\item{max.K}{An array of indices indicating the range where weights are not 0. Indices < max.K at i have weights > 0.}

\item{degree}{A degree of the polynomial of x in the linear regression. The only allowed values are 1 and 2.}
}
\value{
EY.grid
}
\description{
Fits 1D rllm model and generates predictions estimates for each column of a matrix Y
}
