% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magelo.R
\name{mllm.1D}
\alias{mllm.1D}
\title{A local linear 1D model with y being a matrix. It expects bw value}
\usage{
mllm.1D(
  x,
  Y,
  bw,
  y.binary = FALSE,
  with.BB = FALSE,
  grid.size = 400,
  min.K = 15,
  nn.kernel = "epanechnikov"
)
}
\arguments{
\item{x}{A numeric vector of a predictor variable.}

\item{Y}{A matrix such that nrow(Y) = length(x) representation many instances of the outcome variable.}

\item{bw}{A bandwidth parameter.}

\item{y.binary}{A logical variable. If TRUE, bw optimization is going to use a binary loss function mean(y(1-p) + (1-y)p).}

\item{with.BB}{A logical parameter. Set to TRUE if Bayesian bootstraps of the column's mean function are to be returned.}

\item{grid.size}{A number of grid points; was grid.size = 10*length(x), but the results don't seem to be different from 400 which is much faster.}

\item{min.K}{The minimal number of x NN's that must be present in each window.}

\item{nn.kernel}{The name of a kernel that will be applied to NN distances of local linear regression models.}
}
\value{
EY.grid
}
\description{
A local linear 1D model with y being a matrix. It expects bw value
}
