% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{normalize.dist}
\alias{normalize.dist}
\title{Normalizes a distance matrix}
\usage{
normalize.dist(d, min.K, bw)
}
\arguments{
\item{d}{A numeric matrix.}

\item{min.K}{The minimum _\emph{number} of elements of each row of d after normalization with normalized distance < 1.}

\item{bw}{A normalization non-negative constant.}
}
\value{
A list with two components: nd (normalized distances), r (radii).

ALERT: This routine makes sense only in the situation when ncol(d) > minK. Thus,
the user has to make sure this condition is satisfied before calling
normalize_dist().
}
\description{
normalize.dist(d, min.K, bw) divides the rows of the input distance matrix,
d, by 'bw' if the distance of the 'min.K'-th element of the given row, i, is
<bw, if it is not, then all elements are divided by the distance of the
min.K-st element. The radius, r, is bw in the first case and \code{d[i,minK]} in the
second.
}
