% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magelo.R
\name{plot.magelo}
\alias{plot.magelo}
\title{Plot method for magelo objects}
\usage{
\method{plot}{magelo}(
  x,
  type = "fit",
  title = "",
  xlab = "",
  ylab = "",
  with.y.true = TRUE,
  with.pts = FALSE,
  with.CrI = TRUE,
  true.col = "red",
  ma.col = "blue",
  pts.col = "gray60",
  with.predictions.pts = FALSE,
  predictions.pts.col = "blue",
  predictions.pts.pch = 20,
  CrI.as.polygon = TRUE,
  CrI.polygon.col = "gray85",
  CrI.line.col = "gray10",
  CrI.line.lty = 2,
  ylim = NULL,
  legend.cex = 0.8,
  bw = NULL,
  with.legend = TRUE,
  legend.position = "topright",
  legend.inset = 0.05,
  pred.legend.label = "Predictions",
  ...
)
}
\arguments{
\item{x}{An output from magelo()}

\item{type}{Type of plot: "fit", "diagnostic", "residuals" or "residuals.hist"}

\item{title}{Plot title}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{with.y.true}{Logical, whether to plot true values if available}

\item{with.pts}{Logical, whether to plot data points}

\item{with.CrI}{Logical, whether to plot credible intervals}

\item{true.col}{Color for true values}

\item{ma.col}{Color for magelo fit}

\item{pts.col}{Color for data points}

\item{with.predictions.pts}{Logical, whether to show prediction points}

\item{predictions.pts.col}{Color for prediction points}

\item{predictions.pts.pch}{Point character for predictions}

\item{CrI.as.polygon}{Logical, whether to show CrI as polygon}

\item{CrI.polygon.col}{Color for CrI polygon}

\item{CrI.line.col}{Color for CrI lines}

\item{CrI.line.lty}{Line type for CrI lines}

\item{ylim}{Y-axis limits}

\item{legend.cex}{Legend text size}

\item{bw}{Bandwidth value to use (analogous to k in malowess)}

\item{with.legend}{Logical, whether to show legend}

\item{legend.position}{Character string indicating legend position (default: "topright").
One of "bottomright", "bottom", "bottomleft", "left", "topleft",
"top", "topright", "right", or "center"}

\item{legend.inset}{Numeric value for inset distance from the margins as a fraction
of the plot region (default: 0.05)}

\item{pred.legend.label}{Prediction legend label}

\item{...}{Additional parameters passed to plot methods}
}
\description{
Plot method for magelo objects
}
