% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magelo.R
\name{predict.magelo}
\alias{predict.magelo}
\title{Predicting values from a magelo model}
\usage{
\method{predict}{magelo}(object, newdata, ...)
}
\arguments{
\item{object}{an object of class "magelo", typically the result of a call to \code{\link{magelo}}}

\item{newdata}{a numeric vector of values at which predictions are required}

\item{...}{additional arguments (currently unused)}
}
\value{
numeric vector of predicted values
}
\description{
Predicting values from a magelo model
}
\examples{
\dontrun{
model <- magelo(x, y)
predictions <- predict(model, newdata = c(1, 2, 3))
}
}
