% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amagelo.R
\name{print.amagelo}
\alias{print.amagelo}
\title{Print method for amagelo objects}
\usage{
\method{print}{amagelo}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class 'amagelo'}

\item{digits}{Number of digits to display (default: 4)}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for amagelo objects
}
