% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabilo_plus.R
\name{print.mabilo_plus}
\alias{print.mabilo_plus}
\title{Print Method for Mabilo Plus Objects}
\usage{
\method{print}{mabilo_plus}(x, ...)
}
\arguments{
\item{x}{A 'mabilo_plus' object}

\item{...}{Additional arguments (currently unused)}
}
\value{
Returns x invisibly
}
\description{
Prints a concise summary of a mabilo_plus object
}
