% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabilog.R
\name{print.mabilog}
\alias{print.mabilog}
\title{Print Method for Mabilog Objects}
\usage{
\method{print}{mabilog}(x, ...)
}
\arguments{
\item{x}{A 'mabilog' object}

\item{...}{Additional arguments (currently unused)}
}
\value{
Returns x invisibly
}
\description{
Prints a concise summary of a mabilog object
}
