% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwlm.R
\name{print.pwlm}
\alias{print.pwlm}
\title{Print Method for PWLM Objects}
\usage{
\method{print}{pwlm}(x, ...)
}
\arguments{
\item{x}{An object of class "pwlm"}

\item{...}{Additional arguments passed to print}
}
\description{
Print Method for PWLM Objects
}
