% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amagelo.R
\name{print.summary.amagelo}
\alias{print.summary.amagelo}
\title{Print method for summary.amagelo objects}
\usage{
\method{print}{summary.amagelo}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class 'summary.amagelo'}

\item{digits}{Number of digits to display (default: 4)}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for summary.amagelo objects
}
