% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwlm.R
\name{print.summary.pwlm}
\alias{print.summary.pwlm}
\title{Print Method for PWLM Summary Objects}
\usage{
\method{print}{summary.pwlm}(x, ...)
}
\arguments{
\item{x}{An object of class "summary.pwlm"}

\item{...}{Additional arguments passed to print}
}
\description{
Print Method for PWLM Summary Objects
}
