% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magelo.R
\name{rllmf.1D}
\alias{rllmf.1D}
\title{A helper 1D local linear model routine for smoothing predictions.CI's}
\usage{
rllmf.1D(
  x,
  y,
  grid.size = 400,
  degree = 2,
  f = 0.3,
  bw = NULL,
  n.C.itr = 0,
  C = 6,
  stop.C.itr.on.min = TRUE,
  n.cv.folds = 10,
  n.cv.reps = 1,
  n.BB = 0,
  min.K = 5,
  nn.kernel = "epanechnikov"
)
}
\arguments{
\item{x}{A numeric vector of a predictor variable.}

\item{y}{A numeric vector of an outcome variable.}

\item{grid.size}{A number of grid points; was grid.size = 10*length(x), but the
results don't seem to be different from 400 which is much faster.}

\item{degree}{A degree of the polynomial of x in the linear regression; 0
means weighted mean, 1 is regular linear model lm(y ~ x), and deg = d is
lm(y ~ poly(x, d)). The only allowed values are 1 and 2.}

\item{f}{The proportion of the range of x that is used within moving window
to train the model. It can have NULL value in which case the optimal
value of f will be found using minimum median absolute error optimization algorithm.}

\item{bw}{A bandwidth parameter.}

\item{n.C.itr}{The number of Cleveland's absolute residue based reweighting iterations for a robust estimates of mean y values.}

\item{C}{A scaling of |res| parameter changing |res| to |res|/C  before applying ae.kernel to |res|'s.}

\item{stop.C.itr.on.min}{A logical variable, if TRUE, the Cleveland's iterative reweighting stops when the maximum of the absolute values of
differences of the old and new predictions estimates are reach a local minimum.}

\item{n.cv.folds}{The number of cross-validation folds. Used only when f = NULL. Default value: 10.}

\item{n.cv.reps}{The number of repetitions of cross-validation. Used only when f = NULL. Default value: 5.}

\item{n.BB}{The number of Bayesian bootstrap iterations for estimates of CI's of beta's.}

\item{min.K}{The minimal number of x NN's that must be present in each window.}

\item{nn.kernel}{A kernel.}
}
\value{
list of parameters and residues of all linear models
}
\description{
A helper 1D local linear model routine for smoothing predictions.CI's
}
