% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magelo.R
\name{row.eval}
\alias{row.eval}
\title{Row-wise evaluates x at nn.i}
\usage{
row.eval(nn.i, x)
}
\arguments{
\item{nn.i}{An array of indices of K nearest neighbors of the i-th element of x.}

\item{x}{An array of nx elements.}
}
\value{
A matrix obtained from x by evaluating it at the indices of nn.i.
}
\description{
Row-wise evaluates x at nn.i
}
