% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{row.TS.norm}
\alias{row.TS.norm}
\title{Row-wise total sum normalization}
\usage{
row.TS.norm(x)
}
\arguments{
\item{x}{A matrix.}
}
\value{
A matrix obtained from x by dividing the rows of x by the sum of the
given row's elements, if the sum is not 0.
}
\description{
Row-wise total sum normalization
}
