% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{row.weighting}
\alias{row.weighting}
\title{Generates kernel defined weights of the rows of an input numeric matrix}
\usage{
row.weighting(x, bws, kernel.str = "epanechnikov")
}
\arguments{
\item{x}{A numeric matrix.}

\item{bws}{A numeric vector of bandwidths.}

\item{kernel.str}{A character string indicating what kernel to apply to rows of x.

ALERT: Values of max.K can be 0 if there is a row of x, where the weight of
the first elements is 0. Thus \code{max.K[i] = 0} indicates that there are no
non-zero elements in that row.}
}
\description{
Generates kernel defined weights of the rows of an input numeric matrix
}
