% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amagelo.R
\name{summary.amagelo}
\alias{summary.amagelo}
\title{Summary method for amagelo objects}
\usage{
\method{summary}{amagelo}(object, digits = 4, ...)
}
\arguments{
\item{object}{An object of class 'amagelo'}

\item{digits}{Number of digits to display (default: 4)}

\item{...}{Additional arguments passed to summary}
}
\description{
Summary method for amagelo objects
}
