% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwlm.R
\name{summary.pwlm}
\alias{summary.pwlm}
\title{Summary Method for PWLM Objects}
\usage{
\method{summary}{pwlm}(object, ...)
}
\arguments{
\item{object}{An object of class "pwlm"}

\item{...}{Additional arguments passed to summary}
}
\description{
Summary Method for PWLM Objects
}
