% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{tr.exponential.kernel}
\alias{tr.exponential.kernel}
\title{Truncated exponential kernel}
\usage{
tr.exponential.kernel(x, bw = 1)
}
\arguments{
\item{x}{A numeric vector.}

\item{bw}{A bandwidth numeric parameter.}
}
\description{
The function equals to (1-t)*exp(-t) for t within (-1, 1) and 0 otherwise; t = abs(x/bw)
}
