% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{triangular.kernel}
\alias{triangular.kernel}
\title{Triangular kernel}
\usage{
triangular.kernel(x, bw = 1)
}
\arguments{
\item{x}{A numeric vector.}

\item{bw}{A bandwidth numeric parameter.}
}
\description{
The function equals to 1 - abs(x/bw) within (-bw, bw) and 0 otherwise
}
