% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabilo_plus.R
\name{update.mabilo_plus}
\alias{update.mabilo_plus}
\title{Update Mabilo Plus Model}
\usage{
\method{update}{mabilo_plus}(object, ...)
}
\arguments{
\item{object}{A 'mabilo_plus' object}

\item{...}{Arguments to update in the model call}
}
\value{
A new mabilo_plus object
}
\description{
Update a mabilo_plus model with new parameters
}
