% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabilog.R
\name{update.mabilog}
\alias{update.mabilog}
\title{Update Mabilog Model}
\usage{
\method{update}{mabilog}(object, ...)
}
\arguments{
\item{object}{A 'mabilog' object}

\item{...}{Arguments to update in the model call}
}
\value{
A new mabilog object
}
\description{
Update a mabilog model with new parameters
}
