% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabilog.R
\name{vcov.mabilog}
\alias{vcov.mabilog}
\title{Extract Variance-Covariance Matrix}
\usage{
\method{vcov}{mabilog}(object, ...)
}
\arguments{
\item{object}{A 'mabilog' object}

\item{...}{Additional arguments (currently unused)}
}
\value{
Diagonal variance-covariance matrix
}
\description{
Computes an approximate variance-covariance matrix for the predictions
Note: This is a simplified approximation for binomial variance
}
