% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magelo.R
\name{wpredict.1D}
\alias{wpredict.1D}
\title{Predicts the mean values of linear models, beta, over grid points of x}
\usage{
wpredict.1D(beta, nn.i, nn.w, nn.x, nx, max.K, y.binary = FALSE)
}
\arguments{
\item{beta}{Coefficients of the local linear model.}

\item{nn.i}{A matrix of indices of x NN's of xgrid.}

\item{nn.w}{A matrix of NN weights.}

\item{nn.x}{The values of predictor variable over NN's of xgrid.}

\item{nx}{The length of x.}

\item{max.K}{A vector of \strong{indices} indicating the range where weights are not 0.}

\item{y.binary}{Set to TRUE if y is a binary variable.}
}
\description{
It is a variation on predict.1D() with nn.w passed to it instead of nn.d
and nn.kernel for calculation of nn.w inside the corresponding C routine.
}
